+++ processor/src/test/resources/input/com/example/component/GeneratedNonEmptyComponent.java
@@ -1,0 +1,15 @@
+package com.example.component;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+import javax.annotation.Generated;
+
+@Generated( "Some Generator" )
+@ArezComponent( allowEmpty = true )
+public abstract class GeneratedNonEmptyComponent
+{
+  @Action
+  void foo()
+  {
+  }
+}